Remove unneeded if check in UpdateRemoteMirror
UpdateRemoteMirror only attempts to push if there are some refs to push.
Previously there was a simple if check to see if we have something to push.
Later on, batching was added to push in batches of 10 refspecs. The batching
loop loops until there are no refspecs to push anymore. The condition of the
for loop does effectively the same as the if check before it, so let's remove
the unneeded if check to simplify.