blob: Implement pipeline step for git-cat-file(1) batch-checking
In order to unify infrastructure when filtering down a list of revisions
to a list of objects, we're creating a set of pipeline steps which first
enumerate all objects, then extract information about these objects and
finally read the whole objects. This pipeline will allow us to flexibly
put together steps and filter down the object set at intermediate steps.

This commit implements the second pipeline step, which receives as input
a channel of object IDs and outputs information about each of the
objects via `git cat-file --batch-check`.