blob: Add pipeline filters
While we now have a fully functional pipeline in place, it's not yet
able to filter down objects based on the information we got. This is now
implemented via a set of filters which can be inserted as intermediate
steps to reduce the number of objects passed onto the next stage.