blob: Convert `ListLFSPointers()` to use pipeline
Now that the pipelining code is in place, convert `ListLFSPointers()` to
use the new pipeline. The change is currently implemented behind a
feature flag to catch any issues with it, but eventually we can get rid
of the old implementation and only retain the new and much more flexible
pipeline architecture.