blob: Extract sending of LFS pointers via pipelines
Right now, we only have a single RPC which uses the new pipeline code to
enumerate LFS pointers, but we're about to also convert the remaining
ones to use it. These RPCs will require the same logic to send out LFS
pointers as `ListLFSPointers()`.

Extract the code into a separate function to make it reusable and avoid
code duplication.