blackbox: Encapsulate metrics to make them shareable
The metrics for reference discovery as well as for the actual request
are currently hosted directly inside the `Blackbox` structure. We're
about to add a second probe type for pushes which will require the same
set of metrics, which would require us to duplicate code with the
current code layout.

Pull out reference discovery and request metrics into their own
structures to prepare for this and allow us to share the code.