blob: Implement support for object type filters
Implement support for object type filters in the revlist pipeline step.
In combination with `--filter-provided-objects`, this filter will cause
git-rev-list(1) to only print objects which match the filtered type.

This option is not used yet, but will be used in the subsequent commit
which implements object type filters for LFS pointers.