Fix Unix socket address handling following gRPC upgrade
In 29f8bbbc, we upgraded our gRPC library version. The upgrade
introduced some differences in how the Unix socket addresses are handled.
We've previously prefixed unix addresses always with 'unix://'. In the
new version of gRPC, this does not work for relative paths as the first
section following the slashes are interpreted as an authority. Instead,
we should either uses 'unix:relative/path', 'unix:/absolute/path' or
'unix:///absolute/path'. As it is, Gitaly configurations that are using
relative paths for the internal socket directory do not work as it results
in address 'unix://relative/path'. This commit changes the two non-test
locations I found by grepping for 'unix://' to prefix the the socket only
by 'unix:' so both relative and absolute paths work.

Changelog: fixed