Makefile: Silence comments in NOTICE target
In Makefiles, every line of a recipe that's not prefixed by an `@`
character will be printed on invocation of this target. This also
includes shell comments. The NOTICE target has such a set of comments
which are missing the `@`.

Fix this by prefixing those lines with `${Q}` to not have them printed
by default.