blob: Drop object filter from pipeline
As part of our pipeline steps, we have a step which allows one to filter
contents of the last pipeline step. The only additional information that
is available at this point in time is the object contents, so this step
effectively can only filter on contents which have already been read in
the first place. Given that the filters' intent is to avoid loading
objects from disk from which we know that they're not going to be
relevant in the first place, having a filter on object contents is not
really that helpful at all.

More importantly though, we're about to convert the `catfileObject()`
pipeline step to not return object data directly but instead to return a
reader. This is effectively incompatible with the filtering step: we
cannot read the same reader twice, and inspecting object data has been
the only motivation for this filter.

Remove the filter step and instead move the burden of filtering to
callers: they're processing results from `catfileObject()` anyway, so
it's not much of a big deal.