blob: Convert object pipeline step to return an io.Reader
The `catfileObject()` pipeline step used to read the whole object from
disk into memory unconditionally. This wasn't much of a problem while we
only used this code for LFS pointers, which are limited to 200 bytes in
the first place. But we're about to extend the pipeline code such that
we can also use it to serve arbitrarily-sized blobs, which can in fact
be multiple gigabytes in size. It goes without saying that reading
multiple gigabytes into RAM is not something we want to do.

Refactor the pipeline code to return an `io.Reader` interface instead of
slurping in the complete blob contents. This allows the caller to either
read the whole `io.Reader` in case it already knows it's limited (i.e.
in the case of LFS pointers) or to stream contents of the `io.Reader`
via our streamio package.

One design wart is that users of the catfile package mustn't request
another object before the previous object has been fully consumed. But
given that we return the object reader to the caller, we must provide a
means with which we can synchronize having read the complete object and
requesting the next one. This is done via a simple `signallingReader`,
which closes a "done" channel upon seeing `io.EOF`. The pipeline step
will thus wait for the channel to be closed before requesting the next
object.

This is a preparatory step to implement the new `ListBlobs()` RPC in
terms of the pipeline.