Parallel execution of the tests in the commit package
To speed up execution of the tests we mark them allowed to
run in parallel by calling t.Parallel(). As a first step we
mark for parallel execution only upper-level test and no sub-tests.
The tests thar are changing a shared state and revert it back after
test run not marked as those will fail.