blob: Resurrect allocation optimization for LFS pointers
In commit 9f7cb6537 (Reduce memory usage in GetAllLFSPointers,
2021-04-19), a memory allocation optimization was introduced for reading
LFS pointers which avoids having to repeatedly allocate buffers for
objects which aren't LFS pointers in the first place. This optimization
doesn't yet exist in the new pipeline code given that we were always
reading all objects into memory in the pipeline.

Now that the code is returning a reader instead of an allocated object,
let's resurrect the optimization for the pipeline by using the same
pattern with a `bytes.Buffer`.