UserUpdateBranch: Remove Ruby implementation
The RPC has been ported to Go, and now the Ruby implementation can be
removed.

The Sentry interceptor tests were updated, as Sentry tries to 'guess'
based on the call stack what method was called. In this instance I chose
to update it, as pragmattic way forward. Updating this to use a
different RPC was much more work.