stats: Implement support for driving HTTP pushes
Implement support for driving and analyzing HTTP pushes which perform
the equivalent of git-send-pack(1) on a remote repository by specifying
a set of commands stating which references shall be updated along with a
packfile containing all required objects.

Note that in order to test this feature, we have to modify the test HTTP
server to enable anonymous push access.