blackbox: Fix config tests
The blackbox config try to test various failure cases of how the
configuration parser can fail. But because the tests never assert any
specific errors, the tested behaviour in fact differs from the what was
thought to be tested because tests were all missing the Prometheus
listening address.

Fix the issue by asserting more specific error messages and adapting
input such that we actually test what we want to test.