blackbox: Introduce probe type
We're about to introduce a new probe type for HTTP-based pushes. As a
preparatory step, add a new probe type to the blackbox configuration
which allows the admin to discern between these types. For now, only
"fetch" probes are supported -- "push" probes will be added in a
subsequent commit. In order to maintain backwards compatibility with old
configurations, specifying no type will fall back to "fetch" probes.