operations: Drop code supporting UserMergeToRef's precursory ref update
Previous to commit 907c03bf9 (operations: Skip precursory update of
target ref in UserMergeToRef, 2021-06-09), UserMergeToRef used to always
create the target reference even if the merge failed. This behaviour was
unexpected and caused issues for transactions because we have a vote
before an expected failure, which would as a result cause replication
jobs to be created if the merge fails.

This code was replaced by code which simply asserts that the target
reference didn't change while we were computing the merge commit. The
old code continued to exist behind a feature flag though in case this
semantic change were to cause issues.

No issues have been observed, so let's remove the feature flag.