praefect: Avoid double-proxied Gitaly nodes
Some of our Praefect server tests use `testserver.RunGitalyServer()` to
setup their Gitaly nodes. While this isn't a problem by itself, this
function will by default set up a Praefect proxy for the Gitaly nodes if
running the "test-with-praefect" target. The result is that Gitaly nodes
are proxied by two Praefects: first the one set up by above function,
and then the one we're setting up ourselves in the tests. This is not a
supported configuration and may cause RPCs to fail in case Gitaly needs
to reach out to Praefect, e.g. for transactional voting.

Fix the issue by passing the `WithDisablePraefect()` option.