featureflag: Default-enable GoUpdateRemoteMirror
The Go implementation of UpdateRemoteMirror is currently hidden behind a
feature flag. Given that its implementation has been rolled out to
production on June 15th without any further issues, let's default-enable
the feature flag.

This also fixes an issue we have with the rollout of in-memory remotes
in Rails: they're only implemented in the Go port, but due to an
oversight we already default-enabled usage of in-memory remotes in
Rails. As a result, if the GoUpdateRemoteMirror feature flag wasn't
manually enabled, Rails will now always fail.

Changelog: changed