Fix ResolveConflicts RPC not calling Git hooks
Previously the Go implementation of `ResolveConflicts` updated the ref
of a merge request without pushing the changes to the Git hooks. As a
result, GitLab Rails had no idea a branch had been modified, so it could
not refresh associated merge requests as a result.

To support this, we need to centralize the functionality in
update_with_hooks.go so that it can be used by the conflicts package.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/18917

Changelog: fixed