gitpipe: Implement iterator for catfile info step
The pipeline steps are currently easy to misuse: the structures returned
on the channel all have an `Err` field, which MUST be checked by the
caller. It's easy enough to forget if one isn't prepared to do so.

Refactor the catfile info step to instead return a CatfileInfoIterator.
This iterator follows the typical Go idiom of iterators with `Next()`,
`Err()` and `Result()` functions. Given that the iterator is an
interface, this also makes it easily possible for callers to replace it
with their own adapter.