cache: Introduce `Cache` interface
We currently have some issues with flaky tests because of the cache
invalidator creating files after the test has concluded. To fix this, we
need to inject our own cache invalidator in such tests. Prepare for this
change by accepting a new `Cache` interface when constructing servers
instead of using the concrete `DiskCache` structure.