commit: Implement new `ListAllCommits()` RPC
Implement a new `ListAllCommits()` RPC. In contrast to the existing
RPCs, this new RPC will list all commits part of the object database,
regardless of whether the objects are reachable or not. This is done via
git-cat-file(1)'s `--batch-all-objects` switch, which simply enumerates
all objects.

The intent of this RPC is not to list commits of the "main" object
database though, but instead to list commits which have just been
received via git-receive-pack(1). Objects which have been received like
this are first put into a quarantine directory such that they do not end
up in the repository when a push gets rejected. With this new RPC, we
can thus efficiently enumerate all commits part of a push without having
to traverse the object graph.

Changelog: added