commit: Generalize the commits sender
Generalize the implementations of commit senders used by `FindCommits()`
and `FindAllCommits()` such that it can be reused across RPCs. This is
done by accepting a callback function instead of a server stream such
that the caller can simply plug in its own logic to send resulting
objects on the stream.

While at it, this commit also implements a simple change to improve
memory efficiency: instead of reallocating the commits array after each
iteration, we just reset its length to zero. Like this, we reuse the
previously allocated memory.