Require trailing flush in pktline.EachSidebandPacket
This changes the behavior of pktline.EachSidebandPacket to expect a
trailing flush packet. The old behavior was to expect the stream to
end on a packet boundary.

This is part of
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/463, where we
need to be able to signal "end of stream" in a natural way and we
cannot use EOF.

There is only one call site for EachSidebandPacket, and that call site
(PackObjectsHook) only consumes byte streams produced by the same
Gitaly process. In other words, this change is safe in spite of being
incompatible with the old behavior, because it can never be exposed to
the old behavior.

Changelog: other