replicator: Special-case replication of WriteCommitGraph
The WriteCommitGraph RPC is in the same class of idempotent mutators as
Cleanup, GC and similar RPCs: they do modify the repository, but there
shouldn't be any user-visible change in behaviour. As such, it needs to
get special treatment in the replicator so that it doesn't increment
repository generations when created.

Implement this special treatment.

Changelog: performance