catfile: Expose function to parse tags
While the catfile package already implements the logic to parse tags,
this functionality is not exposed to users outside of the package. We'll
need this though in order to implement parsing of commits as part of the
gitpipe code, where we get an `io.Reader` as well as object information.

Refactor the code to pull out a new `ParseTag()` function which parses
raw tags from an `io.Reader`.