gitpipe: Introduce function to transform revision pipelines
While we already have a filter function for the revision pipeline which
allows the caller to purge revision results, a future user will require
the ability to insert additional revisions into the pipeline.

Add a new pipeline step `RevisionTransform()`: given a single item, it
returns a slice of items which replace the current one. This can either
be the item itself in case of an identity mapping, an empty slice to
drop the item or multiple items to add additional revisions into the
pipeline.

Given that this can easily be used to implement `RevisionFilter()`,
this funciton also refactors that function to be implemented in terms of
`RevisionTransform()`.