ListConflictFiles: Allow tree conflicts when specified
In https://gitlab.com/gitlab-org/gitlab/-/issues/281171, we want
to display what type of conflict occurred for a specific file.
There are cases that there will be a missing side (e.g. tree-based
conflicts like renames) and currently ListConflictFiles RPC only
allows returning of conflicts when there is no missing side.

This MR changes ListConflictFiles to accept an AllowTreeConflicts
param which will then be set to true on gitlab-rails when we want
to get the list of conflicts.

Changelog: changed