tempdir: Refactor how new tempdirs are created
Currently, creating new tempdirs is handled mostly in
`newAsRepository()`, which creates a tempdir and then returns both the
tempdir and a Repository protobuf pointing into that tempdir. This is
quite weird given that two of three callers don't actually care about
the repository and just discard it.

Refactor the function into a generic `newDirectory()` function whose
only concern it is to create the tempdir in a specific location.
Assembly of the protobuf is moved into `NewAsRepository()`.