tempdir: Refactor public interface to be more generic
The public interface of the tempdir package is quite focussed on
repositories, even though we don't actually care about the repos
themselves in its implementation. The only information we require is the
storage name of the repository such that we know where to create the
tempdir such that it can be used for any given repo hosted in that
storage.

Refactor public functions to be more generic: all of them now get a
storage name as parameter. Furthermore, `ForDeleteAllRepositories()` is
renamed to `NewWithoutContext()`: the only difference to `New()` is that
it does not take a context and thus doesn't get cleaned up after the
context expires.