tempdir: Fix Goroutine leak in `NewWithoutContext()`
When creating a tempdir via `NewWithoutContext()`, we still spawn the
Goroutine which handles removal of the tempdir when the context expires.
But given that we pass in a `context.Background()` and thus the context
won't ever expire, this Goroutine is essentially stuck forever.

Fix the issue by moving the Goroutine into `New()`. This function is
also called by `NewRepository()`, so both retain the cleanup.