tempdir: Allow synchronizing on tempdir removal
When the context of a tempdir gets cancelled, then we asynchronously
remove the directory. While this is what we want for RPCs given that
it's not useful to block on a potentially lengthy deletion, it does make
testing removal code harder given that we now have to poll.

Fix this issue by adding a channel to the `Dir` structure, which gets
closed when the tempdir was removed. This allows us to easily
synchronize tests with removal of the tempdir and.