updateref: Allow updating references with quarantine directory
The UpdaterWithHooks structure essentially mirrors what happens on a
push: the caller prepares a set of new objects and then tells it which
reference should be updated to point to the new objects. We then go
about and execute the pre-receive, update, reference-transaction and
post-receive hooks in the correct order as it would happen in
git-receive-pack(1). The most important gap though is that we didn't put
objects into a quarantine diretcory like git-receive-pack(1) would. As a
result, even if a hook rejected the update, we'd retain the new objects
in our main object directory.

Now that we have the logic in place to create and migrate quarantine
directories, we can use them to close this gap: if the caller passes us
an object quarantine directory, then we'll call the pre-receive, update
and the first invocation of the reference-transaction hook with the
quarantine environment. Only if they succeed will we now migrate the
objects into the target repository and then call remaining hooks on the
actual repository.

This will allow us both to properly clean up after ourselves in case
hooks refuse an update, and eventually will also allow us to speed
Rails' access checks via the object quarantine directory given that we
can now directly list all new objects via the quarantine directory
instead of having to do a graph walk to determine new objects.