operations: Allow passing quarantine dir to `updateReferenceWithHooks()`
The `updateReferenceWithHooks()` function in the operations package is a
wrapper around `UpdaterWithHooks.UpdateReference()`. It doesn't yet
accept a quarantine directory as parameter and thus always uses over a
`nil` pointer for now. Change that and adapt all callers such that we
can start to make use of quarantine directories in the operations
service.