operations: Split out creation of tags
The `UserCreateTag()` RPC first creates a tag object and then updates
its target reference to point to that new object with hooks. This is two
logical steps, but both are currently implemented in the same function.

Split out creation of the tag into its own function. This will make it
easier in a subsequent commit to pass in a quarantined repository and
see that the tag can only be written into the quarantine object
directory.