Refactor Filesystem into Manager and FilesystemSink
As we are required to support not only a filesystem storage
for the backups current implementation needs to be refactored.
The core logic of backup creation and restoring is untouched
except the operations working with data streams. The new
Sink interface is used to store and retrieve backup data.
Current functionality that allows to store/retrieve repositories
to/from the local filesystem now part of the FilesystemSink
type. And it is used as Sink implementation for the Manager.
The Manager doesn't know anything about underling storage
and just uses methods defined on the interface.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3633