hook: Inject ODB envvars for custom hooks passed via repository proto
When executing custom hooks, we forward the GIT_OBJECT_DIRECTORY and
GIT_ALTERNATE_OBJECT_DIRECTORIES environment variables to the spawned
executables. This is required in the context of object quarantine
directories set up by git-receive-pack(1) such that the child process
knows where to find quarantined objects as well as object of the main
object database.

Until now, we only ever forward these environment variables if the
gitaly-hooks executable had them set up by git-receive-pack(1). This
makes sense given that this was the only place where we actually made
use of quarantine directories. But given that we now also support manual
setup of object quarantine directories for use when updating references
with hooks, this is not sufficient anymore. In this context, we perform
execution of Git hooks via manual invocation of the relevant RPCs, and
thus there is no gitaly-hooks executable involved which would forward
the required environment variables. Instead, information on quarantined
object directories is part of the repository protobuf which gets passed
to us.

Implement the logic to make use of this information and correctly set up
the custom hooks environment in case the repository proto does contain
information about object directories. Like this, custom hooks are now
able to access quarantined objects if executed in the context of a
manually set up quarantine.